+++ actionmailbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Rails 6.0.0.beta1 (January 18, 2019) ##
 
+*   Allow skipping incineration of processed emails.
+
+    This can be done by setting `config.action_mailbox.skip_incineration` to `true`.
+
+    *Pratik Naik*
+
 *   Added to Rails.
 
     *DHH*